Multimedia/requirements.txt
attr==0.3.2
charset_normalizer==3.3.2
checks==0.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.6
Cython==3.0.10
docutils==0.21.2
filelock==3.15.4
HTMLParser==0.0.2
hypothesis==6.108.10
importlib_metadata==8.2.0
ipython==8.12.3
ipython==8.26.0
ipywidgets==8.1.3
Jinja2==3.1.4
jnius==1.1.0
keyring==25.2.1
matplotlib==3.9.0
mock==5.1.0
mtrand==0.1
mypy==1.11.1
pickle5==0.0.12
Pillow==10.4.0
protobuf==5.27.3
psutil==6.0.0
pygame==2.6.0
PyInstaller==6.9.0
pynput==1.7.7
pyOpenSSL==24.1.0
pyOpenSSL==24.2.1
pytest==8.3.2
pytz==2024.1
PyYAML==6.0.1
PyYAML==6.0.1
railroad==0.5.0
redis==5.0.8
setuptools==69.5.1
skimage==0.0
Sphinx==8.0.2
thread==2.0.5
threadpoolctl==3.5.0
tornado==6.4.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
